Add SSH proxy container for development
Because of issues accessing SSH tunnels from Linux add a container
that sets up an SSH tunnel. This requires some set up on the
part of the developer to modify the ~/.ssh/config file, but avoids
having to manually create a tunnel afterwards. From the app's
point of view, it is accessing a normal DB container.